refactor : 자동차들 모음, 자동차 생성 기능 분리 리팩터링
